requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.7.4  # Updated to match langchain requirements
requests>=2.32.5  # Updated to match langchain-community
streamlit>=1.29.0

# AI and ML
google-generativeai>=0.3.1
langchain>=0.3.27
langchain-community>=0.3.29
langchain-core>=0.3.76

# Utils
python-dotenv>=1.0.0
protobuf>=5.0.0,<7.0.0  # Fixed version for opentelemetry-proto
beautifulsoup4>=4.12.0  # For HTML parsing
aiohttp>=3.9.0  # For async HTTP requests

# Optional
ollama>=0.5.3  # If you still need it
pydantic-settings>=2.10.1